refactor(netcdf): extract xarray interop into engines/interop.py (STR-1)
Mirror the dataset/engines collaborator architecture for the NetCDF
god-object (issue #615). Move the bodies of to_xarray/from_xarray and
the private _build_multidim_from_xarray helper out of netcdf.py into a
new netcdf/engines/ subpackage:

- Interop(_Engine) owns to_xarray (instance method).
- from_xarray/_build_multidim_from_xarray become module-level functions
  (from_xarray builds a new container, so it takes cls rather than self).
- NetCDF.to_xarray/from_xarray become thin façades delegating to the
  engine; signatures and return types unchanged.

The engines use distinct attribute names (interop, …) from the eight
inherited Dataset engines so wiring them in __init__ never clobbers an
inherited engine. _update_inplace re-binds them alongside the Dataset
collaborators after the __dict__ swap.
