+++ src/pyramids/netcdf/netcdf.py
@@ -23,7 +23,6 @@ from osgeo import gdal, osr
 from shapely import contains_xy
 
 from pyramids import _io
-from pyramids.base._errors import OptionalPackageDoesNotExist
 from pyramids.base._utils import numpy_to_gdal_dtype
 from pyramids.base.crs import sr_from_epsg
 from pyramids.base.protocols import ArrayLike
@@ -34,6 +33,8 @@ from pyramids.netcdf._kerchunk_facade import combine_kerchunk, to_kerchunk
 from pyramids.netcdf._lazy import apply_unpack, build_lazy_array
 from pyramids.netcdf._mfdataset import open_mfdataset
 from pyramids.netcdf._plot import NetCDFPlot
+from pyramids.netcdf.engines import interop as _interop
+from pyramids.netcdf.engines.interop import Interop
 from pyramids.netcdf.cf import (
     build_coordinate_attrs,
     detect_axis,
@@ -65,6 +66,13 @@ _GDAL_INTEGER_DTYPES = {
 }
 if hasattr(gdal, "GDT_Int8"):
     _GDAL_INTEGER_DTYPES.add(gdal.GDT_Int8)
+
+# Attribute names of the NetCDF-specific engine collaborators (issue #615,
+# STR-1). Kept distinct from the eight inherited `Dataset` engines in
+# `_COLLABORATOR_ATTRS` so wiring them in `NetCDF.__init__` never clobbers an
+# inherited engine. `_update_inplace` re-binds these alongside the Dataset ones
+# after the `__dict__` swap.
+_NETCDF_COLLABORATOR_ATTRS = ("interop",)
 
 
 class _LazyVariableDict(dict):
@@ -536,6 +544,12 @@ class NetCDF(Dataset):
         self._variable_attrs: dict[str, Any] = {}
         self._scale: float | None = None
         self._offset: float | None = None
+        # NetCDF-specific engine collaborators (issue #615, STR-1). Distinct
+        # attribute names from the eight inherited Dataset engines so they do
+        # not clobber `self.io` / `self.spatial` / … . NetCDF exposes thin
+        # façade methods that delegate here (e.g. `nc.to_xarray()` ->
+        # `self.interop.to_xarray()`).
+        self.interop = Interop(self)
 
     def close(self) -> None:
         """Release every GDAL handle this container holds, then close the base.
@@ -636,9 +650,10 @@ class NetCDF(Dataset):
         # so callers using `self.spatial.crop(...)` after this update
         # reach the surviving instance instead of the discarded `new`.
         self_proxy = weakref.proxy(self)
-        # Reuse the canonical collaborator list (single source of truth in dataset.py) so a new
-        # engine — e.g. `georef` — is rebound here too instead of silently keeping a dead back-ref.
-        for attr in _COLLABORATOR_ATTRS:
+        # Reuse the canonical Dataset collaborator list (single source of truth in dataset.py) so a
+        # new engine — e.g. `georef` — is rebound here too instead of silently keeping a dead
+        # back-ref, then rebind the NetCDF-specific engines (`interop`, …) the same way.
+        for attr in (*_COLLABORATOR_ATTRS, *_NETCDF_COLLABORATOR_ATTRS):
             collab = self.__dict__.get(attr)
             if collab is not None:
                 collab._ds = self_proxy
@@ -5724,99 +5739,9 @@ class NetCDF(Dataset):
         self._replace_raster(dst)
         self._invalidate_caches()
 
-    def to_xarray(self) -> Any:
-        """Convert this NetCDF container to an `xarray.Dataset`.
-
-        Builds an in-memory `xarray.Dataset` that mirrors the
-        variables, coordinates, dimensions, and global attributes of
-        this pyramids NetCDF container.
-
-        The entire conversion goes through GDAL's Multidimensional
-        API — the same reader the rest of pyramids' NetCDF code uses.
-        No xarray engine plugin (`netcdf4`, `h5netcdf`,
-        `scipy.io.netcdf`) is involved, so xarray does not need to
-        pull a NetCDF backend: pyramids is the backend. The returned
-        `xr.Dataset` holds already-
-        materialised numpy arrays; for lazy reads use
-        :meth:`read_array(chunks=...)` and wrap the result in
-        :class:`xarray.DataArray` yourself.
-
-        Requires the optional `xarray` package. Install with one of:
-
-        - PyPI: ``pip install xarray``
-        - conda-forge: ``conda install -c conda-forge xarray``
-
-        Returns:
-            xarray.Dataset: An xarray Dataset with the same
-            variables, coordinates, and global attributes.
-
-        Raises:
-            pyramids.base._errors.OptionalPackageDoesNotExist:
-                If `xarray` is not installed.
-            ValueError: If the underlying GDAL handle is not a
-                multidimensional container (open the file with
-                `open_as_multi_dimensional=True`).
-
-        Examples:
-            Convert a pyramids NetCDF to xarray::
-
-                nc = NetCDF.read_file("temperature.nc")
-                ds = nc.to_xarray()
-                print(ds)
-        """
-        try:
-            import xarray as xr
-        except ImportError:
-            raise OptionalPackageDoesNotExist(
-                "xarray is required for to_xarray(). Install with one of:\n"
-                "  - PyPI:        pip install xarray\n"
-                "  - conda-forge: conda install -c conda-forge xarray"
-            )
-
-        rg = self._raster.GetRootGroup()
-        if rg is None:
-            raise ValueError(
-                "to_xarray requires a multidimensional container. "
-                "Open the file with open_as_multi_dimensional=True."
-            )
-
-        coords: dict[str, Any] = {}
-        dims = rg.GetDimensions() or []
-        for d in dims:
-            dim_name = d.GetName()
-            iv = d.GetIndexingVariable()
-            if iv is None:
-                continue
-            coord_attrs = _read_attributes(iv)
-            unit = iv.GetUnit()
-            if unit and "units" not in coord_attrs:
-                coord_attrs["units"] = unit
-            coords[dim_name] = ([dim_name], self._md_array_to_numpy(iv), coord_attrs)
-
-        data_vars: dict[str, Any] = {}
-        for var_name in self.variable_names:
-            md_arr = rg.OpenMDArray(var_name)
-            if md_arr is None:
-                continue
-            arr_dims = md_arr.GetDimensions() or []
-            arr_dim_names = [ad.GetName() for ad in arr_dims]
-            arr_data = self._md_array_to_numpy(md_arr)
-            var_attrs = _read_attributes(md_arr)
-            # GDAL's netCDF driver normalises the CF `units` attribute
-            # to MDArray.GetUnit() / SetUnit() rather than a regular
-            # attribute. Merge it back into var_attrs for a clean
-            # round-trip through xr.Dataset.
-            unit = md_arr.GetUnit()
-            if unit and "units" not in var_attrs:
-                var_attrs["units"] = unit
-            data_vars[var_name] = (arr_dim_names, arr_data, var_attrs)
-
-        result = xr.Dataset(
-            data_vars=data_vars,
-            coords=coords,
-            attrs=self.global_attributes,
-        )
-        return result
+    def to_xarray(self, *args, **kwargs) -> Any:
+        """Facade — delegates to :meth:`Interop.to_xarray <pyramids.netcdf.engines.interop.Interop.to_xarray>`."""
+        return self.interop.to_xarray(*args, **kwargs)
 
     def subset(
         self,
@@ -6506,148 +6431,16 @@ class NetCDF(Dataset):
         dataset: Any,
         path: str | Path | None = None,
     ) -> NetCDF:
-        """Create a pyramids NetCDF from an `xarray.Dataset`.
+        """Facade — delegates to :func:`pyramids.netcdf.engines.interop.from_xarray`.
 
-        Extracts dimensions, coordinates, data variables, and
-        attributes from the `xarray.Dataset` and writes them to a
-        NetCDF file through pyramids' own GDAL Multidimensional
-        writer. No xarray engine plugin (`netcdf4`, `h5netcdf`)
-        is invoked — pyramids is the writer, so xarray does not
-        need to pull a NetCDF backend.
-
-        Usage::
-
-            ds = xr.open_dataset("input.nc")
-            #... xarray processing...
-            nc = NetCDF.from_xarray(ds)
-            var = nc.get_variable("temperature")
-            cropped = var.crop(mask)
-
-        Requires the optional `xarray` package. Install with one of:
-
-        - PyPI: ``pip install xarray``
-        - conda-forge: ``conda install -c conda-forge xarray``
-
-        Args:
-            dataset: An `xarray.Dataset` instance.
-            path: File path where the NetCDF will be written. If
-                `None`, a temp `.nc` is created and cleaned up
-                when the returned object is garbage-collected.
-
-        Returns:
-            NetCDF: A pyramids NetCDF container backed by the data
-            from the xarray Dataset.
-
-        Raises:
-            pyramids.base._errors.OptionalPackageDoesNotExist:
-                If `xarray` is not installed.
-            TypeError: If *dataset* is not an `xarray.Dataset`.
+        See that function for the full contract. ``from_xarray`` builds a new
+        container (it does not operate on an existing instance), so its body
+        lives in the module-level engine function rather than on the
+        instance-bound :class:`~pyramids.netcdf.engines.interop.Interop` engine;
+        ``cls`` is threaded through so the file is read back as the concrete
+        ``Container`` subtype.
         """
-        try:
-            import xarray as xr
-        except ImportError:
-            raise OptionalPackageDoesNotExist(
-                "xarray is required for from_xarray(). Install with one of:\n"
-                "  - PyPI:        pip install xarray\n"
-                "  - conda-forge: conda install -c conda-forge xarray"
-            )
-
-        if not isinstance(dataset, xr.Dataset):
-            raise TypeError(f"Expected xarray.Dataset, got {type(dataset).__name__}")
-
-        cleanup_temp = False
-        if path is not None:
-            path = str(path)
-        else:
-            tmp = tempfile.NamedTemporaryFile(suffix=".nc", delete=False)
-            path = tmp.name
-            tmp.close()
-            cleanup_temp = True
-
-        mem_src = cls._build_multidim_from_xarray(dataset)
-        dst = gdal.GetDriverByName("netCDF").CreateCopy(path, mem_src, 0)
-        if dst is None:
-            raise RuntimeError(f"Failed to write NetCDF to {path}")
-        dst.FlushCache()
-        dst = None
-        mem_src = None
-
-        result = cls.read_file(path, read_only=True)
-        if cleanup_temp:
-            result._xarray_temp_path = path
-            weakref.finalize(result, os.unlink, path)
-        return result
-
-    @staticmethod
-    def _build_multidim_from_xarray(dataset: Any) -> gdal.Dataset:
-        """Build an in-memory GDAL multidim container from an xarray Dataset.
-
-        Creates dimensions from `dataset.sizes`, writes each
-        coordinate as a 1-D indexing MDArray, writes each data
-        variable as an N-D MDArray whose dimensions are resolved by
-        name. Variable and global attributes are copied via pyramids'
-        own `write_attributes_to_md_array` / `write_global_attributes`
-        helpers so every type the CF layer already handles (str, int,
-        float, bool, list) round-trips without going through xarray's
-        NetCDF writer.
-        """
-        src = gdal.GetDriverByName("MEM").CreateMultiDimensional("from_xarray")
-        root = src.GetRootGroup()
-
-        gdal_dims: dict[str, gdal.Dimension] = {}
-        for dim_name, dim_size in dataset.sizes.items():
-            gdal_dims[dim_name] = root.CreateDimension(
-                dim_name,
-                "",
-                "",
-                int(dim_size),
-            )
-
-        def _apply_attrs(md_arr: gdal.MDArray, attrs: dict[str, Any]) -> None:
-            """Write xarray var attrs, routing `units` through SetUnit.
-
-            GDAL's netCDF writer moves the CF `units` attribute onto
-            the MDArray's own unit slot; if we also write it as a regular
-            attribute it's dropped on the next CreateCopy. Split it out
-            so the round trip is lossless.
-            """
-            if not attrs:
-                return
-            remaining = dict(attrs)
-            unit = remaining.pop("units", None)
-            if unit is not None:
-                md_arr.SetUnit(str(unit))
-            if remaining:
-                write_attributes_to_md_array(md_arr, remaining)
-
-        for coord_name, coord in dataset.coords.items():
-            if coord_name not in gdal_dims:
-                continue
-            values = np.asarray(coord.values)
-            ext = gdal.ExtendedDataType.Create(numpy_to_gdal_dtype(values))
-            md_arr = root.CreateMDArray(
-                coord_name,
-                [gdal_dims[coord_name]],
-                ext,
-            )
-            md_arr.Write(np.ascontiguousarray(values))
-            _apply_attrs(md_arr, dict(coord.attrs))
-
-        for var_name, var in dataset.data_vars.items():
-            values = np.asarray(var.values)
-            ext = gdal.ExtendedDataType.Create(numpy_to_gdal_dtype(values))
-            md_arr = root.CreateMDArray(
-                var_name,
-                [gdal_dims[d] for d in var.dims],
-                ext,
-            )
-            md_arr.Write(np.ascontiguousarray(values))
-            _apply_attrs(md_arr, dict(var.attrs))
-
-        if dataset.attrs:
-            write_global_attributes(root, dict(dataset.attrs))
-
-        return src
+        return _interop.from_xarray(cls, dataset, path)
 
 
 class Variable(NetCDF):
